refactoring

+++ src/app/components/modules/song-card/song-card.component.ts
@@ -1,6 +1,8 @@
 import { Clipboard } from '@angular/cdk/clipboard';
-import { Component, EventEmitter, HostListener, Input, OnInit, Output } from '@angular/core';
-import { mergeMap, tap } from 'rxjs/operators';
+import { Component, Input, OnInit } from '@angular/core';
+import { ConfirmationService } from 'primeng/api';
+import { of } from 'rxjs';
+import { catchError, mergeMap, tap } from 'rxjs/operators';
 import { UnsubscribeComponent } from '../../../../models/angular/unsubscribe.model';
 import {
     ECharacteristic,
@@ -10,10 +12,13 @@ import {
 } from '../../../../models/api/api.models';
 import { TInstalled } from '../../../../models/electron/download.model';
 import { TSendDebug, TSendError } from '../../../../models/electron/send.channels';
+import { ILocalMapInfo } from '../../../../models/maps/localMapInfo.model';
+import { TSongHash, TSongId } from '../../../../models/maps/map-ids.model';
 import { ApiHelpers } from '../../../../models/maps/maps.helpers';
 import { LevelStatsData, TLevelStatsInfo } from '../../../../models/player/player-data.model';
+import { ApiService } from '../../../services/null.provided/api.service';
 import { DlService } from '../../../services/null.provided/dl.service';
-import { InstalledSongsService } from '../../../services/null.provided/installed-songs.service';
+import { LocalMapsService } from '../../../services/null.provided/local-maps.service';
 import { PlayerStatsService } from '../../../services/null.provided/player-stats.service';
 import { ElectronService } from '../../../services/root.provided/electron.service';
 import { NotifyService } from '../../../services/root.provided/notify.service';
@@ -23,30 +28,25 @@ import { SongCardService } from './song-card.service';
 @Component({
     selector: 'app-song-card',
     templateUrl: './song-card.component.html',
-    styleUrls: ['./song-card.component.scss']
+    styleUrls: ['./song-card.component.scss'],
+    providers: [ConfirmationService]
 })
 export class SongCardComponent extends UnsubscribeComponent implements OnInit {
-    @HostListener('window:resize', ['$event'])
-    onResize(event: Event) {
-        this.isMobile = window.innerWidth < 992;
-    }
+    @Input() localMode: boolean;
     @Input() tMapDetail?: TMapDetail;
+    @Input() localMapInfo?: ILocalMapInfo;
     public tLevelStatsInfo?: TLevelStatsInfo;
     public isInstalledSong: { status: TInstalled };
     public latestVersion?: TMapVersion;
     public uploadTimeInfo?: string | Date;
 
     private _expanded: boolean;
-    @Input()
     get expanded(): boolean {
-        return this._expanded;
+        return this._expanded && this.tMapDetail != null;
     }
     set expanded(value: boolean) {
         if (value !== this._expanded) this._expanded = value;
     }
-
-    @Output()
-    public expandedChange: EventEmitter<boolean>;
 
     private _diffs?: Map<ECharacteristic, TMapDifficulty[]>;
     get diffs(): Map<ECharacteristic, TMapDifficulty[]> | undefined {
@@ -75,95 +75,65 @@ export class SongCardComponent extends UnsubscribeComponent implements OnInit {
         return this.tMapDetail?.name || 'N/A';
     }
 
-    private _isMobile: boolean;
-    get isMobile(): boolean {
-        return this._isMobile;
-    }
-    set isMobile(val: boolean) {
-        if (val !== this._isMobile) this._isMobile = val;
-    }
-
     get inQueue(): boolean {
         if (this.latestVersion) return this.dlService.has(this.latestVersion);
         else return false;
     }
 
+    get coverUrl(): string {
+        return this.latestVersion?.coverURL || 'assets/bs-default.jpeg';
+    }
+
+    public loading: boolean;
+
     constructor(
         public songPreviewService: SongPreviewService,
         public playerStatsService: PlayerStatsService,
-        public installedSongsService: InstalledSongsService,
+        public installedSongsService: LocalMapsService,
         public dlService: DlService,
         private _eleService: ElectronService,
         private _notify: NotifyService,
         private _clipboard: Clipboard,
-        private _songCardService: SongCardService
+        private _songCardService: SongCardService,
+        private _apiService: ApiService,
+        private _localMapsService: LocalMapsService,
+        private _confirmService: ConfirmationService
     ) {
         super();
         this.isInstalledSong = { status: false };
         this._songNameShort = 'N/A';
-        this.expandedChange = new EventEmitter<boolean>();
         this._expanded = this._songCardService.expandAll;
-        this._isMobile = window.innerWidth < 992;
+        this.localMode = false;
+        this.loading = true;
     }
 
     ngOnInit(): void {
         this.addSub(
             this._songCardService.expandAllChange.pipe(
                 tap((val: boolean) => {
-                    if (this.expanded !== val) this.expanded = val;
+                    if (this._expanded !== val) this.expanded = val;
                 })
             )
         );
-        this.latestVersion = this.tMapDetail?.versions.sort((a: TMapVersion, b: TMapVersion) =>
-            a.createdAt < b.createdAt ? -1 : a.createdAt === b.createdAt ? 0 : 1
-        )[this.tMapDetail.versions.length - 1];
-        if (this.latestVersion != null) {
-            this.diffs = ApiHelpers.getDifficultyGroupedByChar(this.latestVersion);
-            this._setUploadTimeInfo(this.latestVersion.createdAt);
-            this._loadPlayerSongStats()
-                .catch((error: any) => this._eleService.send<TSendError>('ERROR', error))
-                .finally(() => {
-                    this.addSub(
-                        this.playerStatsService.selectedPlayerChange.pipe(
-                            mergeMap(async (result: { name: string } | undefined) => {
-                                try {
-                                    if (result) {
-                                        await this._loadPlayerSongStats();
-                                    }
-                                } catch (error: any) {
-                                    this._eleService.send<TSendError>('ERROR', error);
-                                }
-                            })
-                        )
-                    );
-                    this.addSub(
-                        this.playerStatsService.playerStatsReloaded.pipe(
-                            mergeMap(async () => {
-                                try {
-                                    await this._loadPlayerSongStats();
-                                } catch (error: any) {
-                                    this._eleService.send<TSendError>('ERROR', error);
-                                }
-                            })
-                        )
-                    );
-                });
 
-            this._initIsInstalledSong()
-                .catch(error => this._eleService.send<TSendError>('ERROR', error))
-                .finally(() => {
-                    this.addSub(
-                        this.installedSongsService.installedSongsReloaded.pipe(
-                            mergeMap(async () => {
-                                try {
-                                    await this._initIsInstalledSong();
-                                } catch (error: any) {
-                                    this._eleService.send<TSendError>('ERROR', error);
-                                }
-                            })
-                        )
-                    );
-                });
+        if (!this.localMode && this.tMapDetail) this._init(this.tMapDetail);
+        else if (this.localMapInfo) {
+            const tempTMapDetails = this._localMapsService.getMapDetailCache(this.localMapInfo.id);
+            if (tempTMapDetails) this._init(tempTMapDetails);
+            else {
+                this.addSub(
+                    this._apiService.getById(this.localMapInfo.id).pipe(
+                        tap((tMapDetail: TMapDetail) => {
+                            this._localMapsService.addMapDetailCache(tMapDetail);
+                            this._init(tMapDetail);
+                        }),
+                        catchError(() => {
+                            this._init(undefined);
+                            return of(null);
+                        })
+                    )
+                );
+            }
         }
     }
 
@@ -172,7 +142,7 @@ export class SongCardComponent extends UnsubscribeComponent implements OnInit {
     }
 
     onCardClick() {
-        if (this.tMapDetail && this.latestVersion) {
+        if (!this.localMode && this.tMapDetail && this.latestVersion) {
             if (this.dlService.has(this.latestVersion)) {
                 this.dlService.remove(this.latestVersion);
             } else {
@@ -200,10 +170,87 @@ export class SongCardComponent extends UnsubscribeComponent implements OnInit {
         }
     }
 
-    private async _loadPlayerSongStats(): Promise<void> {
-        if (this.latestVersion?.hash) {
+    async onUninstallSong(event: Event) {
+        this._confirmService.confirm({
+            target: <EventTarget>event.target,
+            message: 'Are you sure that you want to delete this amazing song?',
+            icon: 'pi pi-exclamation-triangle',
+            accept: async () => {
+                try {
+                    const id = this.tMapDetail?.id || this.localMapInfo?.id;
+                    if (id) await this.dlService.deleteSingle(id);
+                } catch (error: any) {
+                    this._notify.error(error);
+                }
+            }
+        });
+    }
+
+    private _init(tMapDetail: TMapDetail | undefined): void {
+        this.tMapDetail = tMapDetail;
+        this.latestVersion = tMapDetail?.versions.sort((a: TMapVersion, b: TMapVersion) =>
+            a.createdAt < b.createdAt ? -1 : a.createdAt === b.createdAt ? 0 : 1
+        )[tMapDetail.versions.length - 1];
+        if (this.latestVersion != null) {
+            this.diffs = ApiHelpers.getDifficultyGroupedByChar(this.latestVersion);
+            this._setUploadTimeInfo(this.latestVersion.createdAt);
+        }
+        const hash = this.latestVersion?.hash || this.localMapInfo?.hash;
+        const id = tMapDetail?.id || this.localMapInfo?.id;
+        if (hash) {
+            this._loadPlayerSongStats(hash)
+                .catch((error: any) => this._eleService.send<TSendError>('ERROR', error))
+                .finally(() => {
+                    this.addSub(
+                        this.playerStatsService.selectedPlayerChange.pipe(
+                            mergeMap(async (result: { name: string } | undefined) => {
+                                try {
+                                    if (result) {
+                                        await this._loadPlayerSongStats(hash);
+                                    }
+                                } catch (error: any) {
+                                    this._eleService.send<TSendError>('ERROR', error);
+                                }
+                            })
+                        )
+                    );
+                    this.addSub(
+                        this.playerStatsService.playerStatsReloaded.pipe(
+                            mergeMap(async () => {
+                                try {
+                                    await this._loadPlayerSongStats(hash);
+                                } catch (error: any) {
+                                    this._eleService.send<TSendError>('ERROR', error);
+                                }
+                            })
+                        )
+                    );
+                });
+        }
+        if (id) {
+            this._initIsInstalledSong(id)
+                .catch(error => this._eleService.send<TSendError>('ERROR', error))
+                .finally(() => {
+                    this.addSub(
+                        this.installedSongsService.installedSongsReloaded.pipe(
+                            mergeMap(async () => {
+                                try {
+                                    await this._initIsInstalledSong(id);
+                                } catch (error: any) {
+                                    this._eleService.send<TSendError>('ERROR', error);
+                                }
+                            })
+                        )
+                    );
+                });
+        }
+        this.loading = false;
+    }
+
+    private async _loadPlayerSongStats(hash: TSongHash): Promise<void> {
+        if (hash) {
             const tLevelStatsInfo = await this.playerStatsService
-                .loadPlayerSongStats(this.latestVersion.hash)
+                .loadPlayerSongStats(hash)
                 .toPromise()
                 .catch(error => {
                     this._eleService.send<TSendError>('ERROR', error);
@@ -224,13 +271,11 @@ export class SongCardComponent extends UnsubscribeComponent implements OnInit {
         }
     }
 
-    private async _initIsInstalledSong(): Promise<void> {
-        if (this.tMapDetail) {
-            const result = await this.installedSongsService
-                .songIsInstalled(this.tMapDetail?.id)
-                .catch(error => {
-                    this._eleService.send<TSendError>('ERROR', error);
-                });
+    private async _initIsInstalledSong(id: TSongId): Promise<void> {
+        if (id) {
+            const result = await this.installedSongsService.songIsInstalled(id).catch(error => {
+                this._eleService.send<TSendError>('ERROR', error);
+            });
             this.isInstalledSong = { status: result && result.result ? 'INSTALLED' : false };
         }
     }
